Make CC3Demo3DTiles demo app compatible with latest code.
Make CC3Demo3DTiles-iOS demo app compatible with latest code.
Add CC3Demo3DTiles-iOS project back into cocos3d-iOS Xcode workspace.
CC3ShaderProgramMatcherBase programForMeshNode: now ensures material
and sets shader program in material. Update README.md.
CC3MeshNode populateFrom: method does not autorelease material.
CC3Node copyChildrenFrom: does not autorelease child copies.
CC3MeshNode ensureMaterial & makeMaterial methods now public and return material.

+++ Projects/CC3Demo3DTiles/CC3Demo3DTiles/Classes/Controls/CCNodeAdornments.h
@@ -103,7 +103,7 @@
  * Initial value is kAdornmentOverZOrder, which places the adornment above the content
  * of the adorned CCNode.
  */
-@property(nonatomic, assign) int zOrder;
+@property(nonatomic, assign) NSInteger zOrder;
 
 /**
  * Most adornments will invoke a CCActionInterval when activated and deactivated in
